requirements.txt
flask==3.0.3
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.2
python-docx==1.1.2
Werkzeug==3.0.4
pypdf==4.3.1
mammoth==1.8.0
playwright==1.44.0
python-dotenv==1.0.1
deep-translator==1.11.4
pymupdf==1.24.5
